requirements.txt
 fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.4.0
openai==1.45.0
anthropic==0.34.2
transformers==4.44.2
torch==2.4.1
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
ray[default]==2.9.3
python-dotenv==1.0.1
aiohttp==3.10.5
tenacity==9.0.0
loguru==0.7.2
rich==13.8.1
matplotlib>=3.8.0
Pillow>=10.0.0
einops>=0.8.0
streamlit>=1.36.0
